https://issues.apache.org/activemq/browse/AMQCPP-272
1. Add the UnsupportedOperationException type to CMS.
2. Update the MessageProducer interface to indicate what exceptions are thrown from the send methods.
3. Fix the ActiveMQProducer class to throw the correct exceptions:
	A. UnsupportedOperationException if called with a NULL Destination and there's not default
	B. UnsupportedOperationException if called with a Destination that is not the default when one was specified.
	C. InvalidDestinatioException if there's a default but Null is passed.
4. Adds new Integration tests to cover the cases outlined above.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/branches/activemq-cpp-3.1.x@890986 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/commands/ActiveMQTextMessage.cpp
@@ -126,8 +126,8 @@ unsigned int ActiveMQTextMessage::getSize() const {
     if( this->text.get() != NULL ) {
         unsigned int size = commands::Message::DEFAULT_MESSAGE_SIZE;
 
-        size += getMarshalledProperties().size();
-        size += this->text->size() * 2;
+        size += (unsigned int)getMarshalledProperties().size();
+        size += (unsigned int)this->text->size() * 2;
 
         return size;
     }
